SDIT-1852 Casenotes migration
Move generic nomis api function
