+++ src/test/kotlin/uk/gov/justice/digital/hmpps/prisonerfromnomismigration/alerts/AlertsNomisApiServiceTest.kt
@@ -245,51 +245,4 @@ class AlertsNomisApiServiceTest {
       assertThat(previousBookingIds.bookingSequence).isEqualTo(2)
     }
   }
-
-  @Nested
-  inner class GetPrisonerIds {
-    @Test
-    internal fun `will pass oath2 token to service`() = runTest {
-      alertsNomisApiMockServer.stubGetPrisonIds()
-
-      apiService.getPrisonerIds(
-        pageNumber = 0,
-        pageSize = 20,
-      )
-
-      alertsNomisApiMockServer.verify(
-        getRequestedFor(anyUrl()).withHeader("Authorization", equalTo("Bearer ABCDE")),
-      )
-    }
-
-    @Test
-    internal fun `will pass page params to service`() = runTest {
-      alertsNomisApiMockServer.stubGetPrisonIds()
-
-      apiService.getPrisonerIds(
-        pageNumber = 5,
-        pageSize = 100,
-      )
-
-      alertsNomisApiMockServer.verify(
-        getRequestedFor(urlPathEqualTo("/prisoners/ids/all"))
-          .withQueryParam("page", equalTo("5"))
-          .withQueryParam("size", equalTo("100")),
-      )
-    }
-
-    @Test
-    fun `will return a page of alerts`() = runTest {
-      alertsNomisApiMockServer.stubGetPrisonIds(totalElements = 10, offenderNo = "A0001KT")
-
-      val prisonerIds = apiService.getPrisonerIds(
-        pageNumber = 5,
-        pageSize = 100,
-      )
-
-      assertThat(prisonerIds.content).hasSize(10)
-      assertThat(prisonerIds.content[0].offenderNo).isEqualTo("A0001KT")
-      assertThat(prisonerIds.content[1].offenderNo).isEqualTo("A0002KT")
-    }
-  }
 }
